Don't check compiled assets into git
If the files are present in git, then Heroku will try to use them
instead of compiling the assets itself. There are more details here:
https://devcenter.heroku.com/articles/rails-asset-pipeline

But in the end, it's simpler to just not check the files into git, and
let Heroku re-compile the assets when it needs to (whenever you push to
Heroku).